Refactored some code to a helper for unit tests.
Fixed age rating not reflecting changes at series detail page.
